+++ test/onyx/peer/min_peers_test.clj
@@ -63,6 +63,7 @@
                         {:lifecycle/task :out
                          :lifecycle/calls :onyx.plugin.core-async/writer-calls}]
             _ (reset! in-chan (chan (inc n-messages)))
+            _ (reset! out-chan (chan (sliding-buffer (inc n-messages))))
             _ (doseq [n (range n-messages)]
                 (>!! @in-chan {:n n}))
             _ (>!! @in-chan :done)
@@ -72,7 +73,6 @@
                                     :workflow workflow
                                     :lifecycles lifecycles
                                     :task-scheduler :onyx.task-scheduler/balanced})
-            _ (reset! out-chan (chan (sliding-buffer (inc n-messages))))
             results (take-segments! @out-chan)]
 
         (let [expected (set (map (fn [x] {:n (inc x)}) (range n-messages)))]